Add customizable modal
By using the menuModal you can easily customize a modal that fits your needs. You only need to include the menuModal.ejs partial file and the menuModal.js file. Now you can change the values of the menuModal object to reflect the elements you want in the modal. After that you can call the createModal() function to display it