Add line drawing with collision detection and holes
The line moves at a constant speed and you can change the direction of the line by pressing left or right on your keyboard. Collision detection has been added so if it collides with a wall or its trail then it's game over. I also added semi randomly generated holes in the line which doesn't have collision