Add dynamic player info on leaderboard and chat
When you connect to a game you'll receive the game state containing all the players in the room. This info is used to generate the dynamic leaderboard showing all the players in the room and their associated colors. When a new player joins, a message is emitted to the other users containing the new player whereafter they get added to the leaderboard. If a user disconnects from the websocket the specific user is then also removed from the leaderboard. The chat names also displays the color associated to that user now