Add multiplayer functionality
All the calculations for the placement of the line is now handled on the server. To make the server calculate the coordinates for the next placement you need to send the keystate to the socket event listener called: 'updatePosition'. I've imlemented an interval loop that calls this event 60 times per second and ones the player collides with another player or a wall the interval is cleared and the line path is deleted from the canvas and player object. Inside the gameClasses file i've made functions with the suffix: 'DTO' which stands for Data Transfer Object. Essentially you don't need to send the whole player class since theres a lot of information that is only relevant to the server. Therefore you can use these DTO methods to make a smaller copy of the player instance that only includes what is relevant to the client. I've also added flying functionality which is basically the same as jump but it lasts longer and its currently only applied when the line is beginning to move