Optimize data transfer object
There is now a function called playerPosDTO that doesn't include the path of the player since it uses a large amount of bytes and can potentially cause network issues. The current position of the player is instead send to the client where we push the value into the path stored on the client. To reduce problems with this method we synchronize with the server every other second to ensure we have the right data on the client