Improve collision detection
Previously we had a value called 'Buffer' that was initially used to not check the last points in the current player path so it didn't collide with itself instantly. This value didn't work with spatial hashing and was only intended for the previous linear way of checking for collisions. Now it correctly checks if the player has collided with its recent points in path. We check if the position we collided with in the spatial hash is equal to a point in the recent player path that is retreived by using slice. The amount of values retreived from the player path is equal to the line width size